+++ app/Http/Controllers/Api/EventController.php
@@ -7,6 +7,7 @@ use App\Http\Resources\EventResource;
 use App\Http\Traits\CanLoadRelationships;
 use App\Models\Event;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class EventController extends Controller
 {
@@ -18,6 +19,7 @@ class EventController extends Controller
      */
     public function index()
     {
+        Gate::authorize('viewAny', Event::class);
         $query = $this->loadRelationships(Event::query());
 
         return EventResource::collection(
@@ -32,14 +34,15 @@ class EventController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('create', Event::class);
         $event = Event::create([
             ...$request->validate([
-                "name" => "required|string|max:255",
-                "description" => "nullable|string",
-                "start_time" => "required|date",
-                "end_time" => "required|date|after:start_time",
+                'name' => 'required|string|max:255',
+                'description' => 'nullable|string',
+                'start_time' => 'required|date',
+                'end_time' => 'required|date|after:start_time'
             ]),
-            'user_id' => 1
+            'user_id' => $request->user()->id
         ]);
 
         return new EventResource($this->loadRelationships($event));
@@ -50,7 +53,10 @@ class EventController extends Controller
      */
     public function show(Event $event)
     {
-        return new EventResource($this->loadRelationships($event));
+        Gate::authorize('view', $event);
+        return new EventResource(
+            $this->loadRelationships($event)
+        );
     }
 
     /**
@@ -58,12 +64,13 @@ class EventController extends Controller
      */
     public function update(Request $request, Event $event)
     {
+        Gate::authorize('update', $event);
         $event->update(
             $request->validate([
-                "name" => "sometimes|string|max:255",
-                "description" => "nullable|string",
-                "start_time" => "sometimes|date",
-                "end_time" => "sometimes|date|after:start_time",
+                'name' => 'sometimes|string|max:255',
+                'description' => 'nullable|string',
+                'start_time' => 'sometimes|date',
+                'end_time' => 'sometimes|date|after:start_time'
             ])
         );
 
@@ -75,10 +82,9 @@ class EventController extends Controller
      */
     public function destroy(Event $event)
     {
+        Gate::authorize('delete', $event);
         $event->delete();
 
-        return response()->json([
-            'message' => 'Event deleted successfully',
-        ]);
+        return response(status: 204);
     }
 }
